Provide way to migrate to augeasproviders_sysctl
duritong/sysctl [1] is a very old module and not supported on forge. We need
to move to augeasproviders_sysctl [2] which is the only approved sysctl
module on forge now. More information can be found here [3].

To do that with backwards compatibility, we renamed the type and
provider for duritong/sysctl. There is only caller for sysctl type at
sysctl::value, so this is pretty safe.

This means the new augeasproviders_sysctl will take the sysctl resource
type name, and old duritong/sysctl will use the duritongsysctl resource
type name.

Also added a deprecation notice and link to wiki change to aid
transition.

Note: `puppet generate types` need to be run after this change and
installing augeasproviders_sysctl to update the pointers.

[1] https: //forge.puppet.com/modules/duritong/sysctl
[2] https: //forge.puppet.com/modules/herculesteam/augeasproviders_sysctl
[3] https: //trello.com/c/gqox1JMi/1970-migrate-puppet-sysctl
Change-Id: Iac04bfa54a4f350c7e028351b20be225cd19f8dc